Add bullseye template.
Comparing against the one shipped in bullseye didn't show any differences from
the Buster template.

However, I did notice that we *hard-code* one option, but it is also parameterized:

GSSAPIAuthentication <%= scope.lookupvar('::sshd::gssapi_authentication') %>

and then later in the template is:

GSSAPIAuthentication no

I removed this hard-coded option.

Additionally, the "UseDNS no" option seems to be default set to this option, and
is commented earlier in the file, so this is an unnecessary setting.