add support for the Include directory
Right now we only add it to the Debian bullseye configuration, as I'm
not sure earlier versions support it correctly, and I'm not familiar
with the other distros to make a call.

But we *do* start managing it as a File resource, which means
configurations that were manually added there will be dropped, unless
an associated File resource is created of course.