allow overriding host aliases exported for the SSH key
I have a host that is available as (say) marcos.anarc.at,
and shell.anarc.at. The latter gives a host key error because it's not
exported (and there's no obvious way to do so in the module right
now).

Allow the user to override this, by setting a host_aliases class
variable in the base class, defaulting to the FQDN. This can be
extended with (say) this in Hiera:

sshd::host_aliases:
  - "%{facts.fqdn}"
  - "smtp.anarc.at"