remove duplicate SSH key export
This class looks like dead code. The top level comment seems to be
incomplete, and it duplicates code already present in
client/base.pp. It's unused anywhere.

If we do want to have a mechanism to do only this export, we should
reuse it in client/base.pp, because as it stands it has a few bugs
including missing ECC keys and possibly others.