ssh_keygen: remove hardcoded bit size
This is necessary to support other key types than RSA, which I
consider much higher priority than switching away from the
default (3072 in bookworm, vs 4096 before this commit).

This could be re-established later, but then we'd have to keep a
mapping between the key types and bit sizes, which seems rather
perilous.