host_aliases is an Array
In #41, we implemented support for the host_aliases paramater to the
ssh_key type. But that parameter actually expects an array, not a
string, or at least it *supports* an array (depending on the
version?).

In any case, we certainly want this to be an array, and it's how I
currently do this in production.