handle SSH bootstrap correctly
Before the merge with ng, we would just check `::sshrsakey` or
`$facts['sshrsakey']` which can't fail if it's empty (it's just
empty). But now it's actually possible that there's no
`$facts['ssh']`. For example, if SSH is not installed yet (which was
my use case when bootstrapping this laptop), there's just no SSH key
at all, and the builtin  fact doesn't populate that array.

The error i'd get is:

    Operator '[]' is not applicable to an Undef Value.

... on base.pp line 47, column 14 (which is the `f` in `$facts`, so
not every telling). It's actually possible that `$facts['ssh']` is
defined and not `$facts['ssh']['rsa']` as well, which is why we go
through those convoluted checks.

This not ideal: we should instead iterate over the array and just
export whatever key is in there, possibly. But for now this scratches
my own itch and should fix the bootstrap problem.